+++ examples/playground/src/main.ts
@@ -1,6 +1,7 @@
 import { createApp } from 'vue'
 import ElementPlus from 'element-plus'
 import { G2rainUi } from '@g2rain/ui'
+import { G2rainPlatformUi } from '@g2rain/ui/platform'
 import 'element-plus/dist/index.css'
 import '@g2rain/theme/styles.css'
 import '@g2rain/ui/style.css'
@@ -11,6 +12,8 @@ createApp(App)
   .use(ElementPlus)
   .use(G2rainUi, {
     translate: (_key, fallback) => fallback,
+  })
+  .use(G2rainPlatformUi, {
     dataProviders: {
       organ: {
         loadOptions: async ({ key, value }) => [