+++ packages/runtime/src/index.ts
@@ -1,5 +1,0 @@
-export * from './loading/index'
-export * from './micro-app/index'
-export * from './permission/index'
-export * from './theme/index'
-