+++ packages/ui/src/remote-select/index.ts
@@ -1,6 +1,5 @@
 export { default as RemoteSelect } from './RemoteSelect.vue'
 export { default as ApiSelect } from './ApiSelect.vue'
-export { default as DictSelect } from './DictSelect.vue'
 export type {
   FetchDataFunction,
   RemoteSelectExpose,